report/requirements.txt
pandas>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
